src/mfi_ddb/retrieval_api/requirements.txt
FastAPI[all]
psycopg2-binary
dateutils